backend/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
python-dotenv==0.21.0
openai==0.27.0
requests==2.28.1
pydantic==1.10.2
python-multipart==0.0.5 
pymysql==1.0.2
cryptography==3.3.2
